separate out dwells
